+++ src/main/kotlin/study/racinggame/domain/RaceRule.kt
@@ -1,4 +1,4 @@
-package study.racinggame.service
+package study.racinggame.domain
 
 import kotlin.random.Random
