+++ src/main/kotlin/study/racinggame/reader/ActualInputReader.kt
@@ -1,4 +1,4 @@
-package study.racinggame.service.reader
+package study.racinggame.reader
 
 class ActualInputReader : InputReader {
     override fun raedLine() = readln()
